requirements.txt
pandas==2.2.2
numpy==1.26.4
pyarrow==15.0.2
scikit-learn==1.4.2
lightgbm==4.3.0
catboost==1.2.5
optuna==3.6.1
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
matplotlib==3.8.4
seaborn==0.13.2
ydata-profiling==4.8.3
plotly==5.21.0
pyyaml==6.0.1
charset-normalizer==3.3.2
joblib==1.4.2
scipy==1.12.0
tqdm==4.66.2
mlflow==2.12.1
requests==2.31.0
rich==13.7.1
